Fill in stacktrace_targets during GitHubSync
We can backfill this data from issue comments, since later comments will override the value for the current target of
the stacktrace just as they do when the comments are received from the webhook. This should also allow us to backfill
the entire database contents from the GitHub issues directly, which is a nice added benefit.